+++ tests/Unit/AddMultyFilterRuleValidation.php
@@ -19,7 +19,7 @@ it('can add filter rule using pools callback', function () {
     $valid = new Validator(['test' => ' test ', 'test2' => ' test ']);
 
     $valid->field('test', 'test2')->required();
-    $valid->filter_pool(fn (FilterPool $f) => [
+    $valid->filters(fn (FilterPool $f) => [
         $f->test->trim(),
         $f->test2->trim(),
     ]);